This is a model designed to be mounted on the GR86 flap.
The case is compatible with both the Adafruit Feather nRF52 (3406) and the Adafruit ESP32-S3 Feather 8MB/0MB PSRAM (5323).
It is optimized for the MLX90640 with a 110° field of view. When installing the temperature sensor, it is intended to be used together with the Temperature Sensor Mounting Spacer 3mm for MLX90640.